Refactor DCModelComp imports and add UV installation in CI.
Updated import paths for `DCModelComp` to reflect the new module structure, aligning with the `dc` package. Additionally, included a step in the CI workflow to install UV dependencies for Python, improving testing consistency.
