+++ tests/comp/mljar/conftest.py
@@ -5,7 +5,7 @@ import pytest
 from ml_orchestrator import artifacts
 
 from tabular_orchestrated.components import DataSplitter
-from tabular_orchestrated.deepchecks import DCModelComp
+from tabular_orchestrated.dc import DCModelComp
 from tabular_orchestrated.mljar.mljar import EvaluateMLJAR, MLJARTraining
 from tabular_orchestrated.mljar.mljar_deepchecks import MljarDCModelComp
