+++ tests/comp/mljar/test_mljar.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from ml_orchestrator import artifacts
 from sklearn import datasets
 
-from tabular_orchestrated.deepchecks import DCModelComp
+from tabular_orchestrated.dc import DCModelComp
 from tabular_orchestrated.mljar.mljar import EvaluateMLJAR, MLJARTraining
 
 dataset_importers = [
